+++ internal/cloudflare/sync.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/3011/cfscan/internal/model"
+	"github.com/3011/cfscan/v2/internal/model"
 )
 
 const SourceOfficial = "cloudflare_official"