+++ internal/cloudflare/sync_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/3011/cfscan/internal/model"
+	"github.com/3011/cfscan/v2/internal/model"
 )
 
 func TestFetchASNPrefixes(t *testing.T) {