+++ internal/store/postgres/results_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/3011/cfscan/internal/model"
+	"github.com/3011/cfscan/v2/internal/model"
 )
 
 func TestLatestResultDatasetFiltersStatusAfterRanking(t *testing.T) {