+++ internal/store/postgres/store.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/3011/cfscan/internal/model"
+	"github.com/3011/cfscan/v2/internal/model"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 )